Allow suppressing embeds on edits too
