+++ src/main/java/net/dv8tion/jda/api/utils/messages/MessageCreateBuilder.java
@@ -17,7 +17,6 @@
 package net.dv8tion.jda.api.utils.messages;
 
 import net.dv8tion.jda.api.entities.Message;
-import net.dv8tion.jda.api.entities.Message.MessageFlag;
 import net.dv8tion.jda.api.entities.MessageEmbed;
 import net.dv8tion.jda.api.interactions.components.LayoutComponent;
 import net.dv8tion.jda.api.utils.AllowedMentions;
@@ -209,24 +208,6 @@ public class MessageCreateBuilder extends AbstractMessageBuilder<MessageCreateDa
     {
         this.tts = tts;
         return this;
-    }
-
-    @Nonnull
-    @Override
-    public MessageCreateBuilder setSuppressEmbeds(boolean suppress)
-    {
-        int flag = MessageFlag.EMBEDS_SUPPRESSED.getValue();
-        if (suppress)
-            this.flags |= flag;
-        else
-            this.flags &= ~flag;
-        return this;
-    }
-
-    @Override
-    public boolean isSuppressEmbeds()
-    {
-        return (this.flags & MessageFlag.EMBEDS_SUPPRESSED.getValue()) != 0;
     }
 
     @Override
