+++ src/main/java/net/dv8tion/jda/internal/utils/message/MessageCreateBuilderMixin.java
@@ -72,20 +72,6 @@ public interface MessageCreateBuilderMixin<R extends MessageCreateRequest<R>> ex
 
     @Nonnull
     @Override
-    default R setSuppressEmbeds(boolean suppress)
-    {
-        getBuilder().setSuppressEmbeds(suppress);
-        return (R) this;
-    }
-
-    @Override
-    default boolean isSuppressEmbeds()
-    {
-        return getBuilder().isSuppressEmbeds();
-    }
-
-    @Nonnull
-    @Override
     default R setFiles(@Nullable Collection<? extends FileUpload> files)
     {
         getBuilder().setFiles(files);
